api/requirements.txt
asgiref==3.9.2
Django==5.2.6
psycopg2==2.9.10
tzdata==2025.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.0
django-filter==24.3
django-cors-headers==4.3.1